requirements.txt
azure-storage-blob
azure-identity
pandas
sqlalchemy
pyodbc
streamlit
dotenv
